+++ src/coreclr/debug/createdump/crashinfo.cpp
@@ -122,10 +122,13 @@ CrashInfo::GatherCrashInfo(MINIDUMP_TYPE minidumpType)
         return false;
     }
 #endif
-    TRACE("Module addresses:\n");
-    for (const MemoryRegion& region : m_moduleAddresses)
+    if (g_diagnosticsVerbose)
     {
-        region.Trace();
+        TRACE_VERBOSE("Module addresses:\n");
+        for (const MemoryRegion& region : m_moduleAddresses)
+        {
+            region.Trace();
+        }
     }
     // If full memory dump, include everything regardless of permissions
     if (minidumpType & MiniDumpWithFullMemory)
@@ -589,7 +592,7 @@ CrashInfo::CombineMemoryRegions()
 
     TRACE("CombineMemoryRegions: FINISHED\n");
 
-    if (g_diagnostics)
+    if (g_diagnosticsVerbose)
     {
         TRACE("Memory Regions:\n");
         for (const MemoryRegion& region : m_memoryRegions)
@@ -619,7 +622,21 @@ CrashInfo::SearchMemoryRegions(const std::set<MemoryRegion>& regions, const Memo
 void
 CrashInfo::Trace(const char* format, ...)
 {
-    if (g_diagnostics) {
+    if (g_diagnostics)
+    {
+        va_list args;
+        va_start(args, format);
+        vfprintf(stdout, format, args);
+        fflush(stdout);
+        va_end(args);
+    }
+}
+
+void
+CrashInfo::TraceVerbose(const char* format, ...)
+{
+    if (g_diagnosticsVerbose)
+    {
         va_list args;
         va_start(args, format);
         vfprintf(stdout, format, args);
