+++ src/coreclr/debug/dbgutil/elfreader.h
@@ -66,4 +66,5 @@ private:
     virtual void VisitProgramHeader(uint64_t loadbias, uint64_t baseAddress, ElfW(Phdr)* phdr) { };
     virtual bool ReadMemory(void* address, void* buffer, size_t size) = 0;
     virtual void Trace(const char* format, ...) { };
+    virtual void TraceVerbose(const char* format, ...) { };
 };
